requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Graph Neural Networks
torch-geometric>=2.3.0
networkx>=3.1

# Vector Database
chromadb>=0.4.0

# LLM Integration (Vertex AI)
langchain>=0.0.350
langchain-google-vertexai>=1.0.0
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Baselines and Evaluation
rank-bm25>=0.2.2
ragas>=0.1.0

# Web Interface
flask>=3.0.0
